Use hackage releases of libs, and downgrade to lts-13.26
Use the new hackage releases of cborg and canonical-json.

Downgrade from stackage lts-13.28 to lts-13.26. No dependent packages
were actually using this because the nix setup isn't ready for it yet.

Downgrading to match reality makes it easier for our tools
(cardano-repo-tool) to generate the right info. The tool uses the
stackage snapshot from this repo for the generated snapshot for all
other repos. So it's really helpful if this one is a working one,
otherwise the generated stuff has to be edited by hand.