Update gitignore; cabal.project; aeson-pretty version bounds
`aeson-pretty`'s Config datatype was updated in 0.8.5 to have a 4th
field. This wasn't captured in the cabal file and I ran into a build
error in a downstream package when an older version of `aeson-pretty`
was built which lacked the 4th field.

I also add /test to the `cabal.project` and gitignore the GHC env file.