Make GC optional
GC is expensive, and so we should skip if it a slight over-estimation of
the count is acceptable. This however makes it more important that we
avoid GC moving the object to be counted after taking its address, so
use a stable pointer for that (we should probably have done that from
the start).