Remove cardano-crypto package from snapshot.yaml
`cardano-crypto` is not used in this project.

Downstream packages using this snapshot.yaml as their resolver automatically
get this `cardano-crypto` version pin, *but only for Stack*. The
`cabal.project` of these repositories must then contain a version pin for
`cardano-crypto`. This means that the hashes of these two are likely to get
out of sync (one is defined in cardano-prelude and the other in the repository
in question, e.g., ouroboros-network). Moreover, it's inconsistent for
repositories to have a version pin in `cabal.project` but not in `stack.yaml`.

A breaking, but intentional consequence of this change is that downstream
repositories will need to add a version pin on cardano-crypto to their
`stack.yaml`.