fix: memory leak in Hasura GraphQL client
The ApolloClient stack appears to retain query variables, causing growth in the heap.
It has become apparent since implementing the asset batch operations. This commit
replaces the package with a simplified client from graphql-request. ApolloClient is still
used in tests for the purpose of utilising the Apollo integration test server.