fix: asset metadata server fetch regression
A bug was introduced in #526, and the test cases for this feature are not sufficient to produce a fail.
This commit refactors the implementation to contain the hex-prefixing to HasuraClient,
and therefore ensures the metadata server request contains the assetId in the format required. A concrete
test assertion has been added.