feature: adds scripts to simplify up and down commands
The native Docker commands along with the Docker Buildkit host config are a
little verbose, and do not present an amazing interface for end-users. This
commit adds a convenience script to simplify the startup and teardown of the
stack with the docker-compose ENVs exposed. The use of the cache is opt-in,
but is used in the documented invocation.