refactor: upgrade to Hasura 2 to move Asset table to separate DB
This refactor improves management of the data created by Cardano GraphQL.
By isolating it from the DB managed by `cardano-db-sync`, there's better separation to assist
with independent snapshots, and rebuilding becomes more streamlined since it's rarely needed
for the asset table given the stability of that model. To achieve this Hasura 2 is needed for
cross-DB relationships. As a side-effect, it will likely address some known issues with memory
leaks and other performance improvements.